Finished config update, onPlayerMove handler
Updated the portal vector configuration handling so now it reads and
writes vector lists to describe portal shapes, added onPlayerMove
handler that teleports players in a portal.

Need to add a toggle or something to prevent unintentional
re-teleportation when exiting the other portal and to add the actual
portal geometry handling methods.