Completed the update to the configuration handling to allow reading/writing of vector lists to describe portal geometry. Also added an onPlayerMove method to do the actual teleporting. Will need to develop the portal geometry methods as well as find a method to prevent return-teleports when trying to exit a teleport.Updated config, added onPlayerMove handler
Completed the update to the configuration handling to allow
reading/writing of vector lists to describe portal geometry. Also added an
onPlayerMove method to do the actual teleporting. Will need to develop the
portal geometry methods as well as find a method to prevent
return-teleports when trying to exit a teleport.