Updated onBlockPhysics handler
Now the handler checks if the changed block was of the frame material
type and if so, runs a full portal integrity (new Portals class member)
check. This iterates through every portal frame block configured, checks
that it is still of the frame material type and if not, destroys the
associated portal.