Bug fix - Minecarts
I totally forgot to test behavior when a player is riding a minecart
through a portal... Turns out that their Y (altitude) is measured as 1m
lower than actualwhen riding a card, which meant that they can be in the
portal but be registered as being below it. Added a test to see if they
are riding a minecart, add 1m to their Y and pass that modified that
location to necessary members to allow for proper calculation of
locations.