Added check for solid blocks at destination portal
Now, if the default North/West face of the portal has more solid blocks
in front of it than the South/East face, the player's destination yaw is
set so they are facing the less-solid portal opening. This way, players
won't be teleported facing a wall if the portal is build on against a
solid wall.